+++ source/engine/CoreSystems/include/CoreSystems/Geometry/Indices/IBowIndicesBase.h
@@ -20,7 +20,7 @@ namespace bow {
 	public:
 		virtual ~IIndicesBase(){}
 
-		virtual unsigned int Size() = 0;
+		virtual size_t Size() = 0;
 
 	public:
 		const IndicesType Type;
